requirements.txt
absl-py==1.4.0
astroid==2.15.6
asttokens==2.2.1
attrs==22.2.0
autopep8==2.0.4
backcall==0.2.0
backports.functools-lru-cache==1.6.5
cachetools==5.3.1
carla==0.9.10
certifi==2023.7.22
charset-normalizer==3.2.0
cloudpickle==2.2.1
cmake==3.27.2
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.7
discrete-signals==0.8.3
entrypoints==0.4
exceptiongroup==1.2.0
executing==1.2.0
Farama-Notifications==0.0.4
filelock==3.12.2
fonttools==4.42.0
funcy==1.18
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
gymnasium==0.28.1
highway-env==1.8.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
ipykernel==6.25.1
ipython==8.14.0
isort==5.12.0
jax-jumpy==1.0.0
jedi==0.19.0
Jinja2==3.1.2
jupyter-client==7.3.4
jupyter_core==5.3.1
kiwisolver==1.4.4
lark-parser==0.12.0
lazy-object-proxy==1.9.0
lenses==0.5.0
lit==16.0.6
ltlf2dfa==1.0.2
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mccabe==0.7.0
metric-temporal-logic==0.4.1
mpmath==1.3.0
mypy==0.981
mypy-extensions==0.4.3
nest-asyncio==1.5.6
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging==23.1
pandas==1.3.5
parsimonious==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==23.2.1
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pydot==1.4.2
pygame==2.5.0
Pygments==2.16.1
pylint==2.17.5
pyparsing==3.0.9
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
rustworkx==0.13.1
sb3-contrib==2.0.0a4
scipy==1.11.1
setuptools==68.0.0
shapely==2.0.1
singledispatch==4.1.0
six==1.16.0
sortedcontainers==2.4.0
stable-baselines3==2.0.0
stack-data==0.6.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
torch==2.0.1
tornado==6.1
tqdm==4.66.1
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.7
wheel==0.38.4
wrapt==1.15.0
zipp==3.16.2
